+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-# 1.47.0
+## 1.48.0
+
+- Added `\Wizaplace\SDK\Order\Order::getShippingCost`
+- Added `\Wizaplace\SDK\Order\Order::getDiscount`
+
+## 1.47.0
 
 ## 1.46.0